Fix slow dask processing
Dask chaches results and this is why the code from yestday was so fast.
Trying again today didn't give the same result as the cache was empty.
Looking again at the dashboard I realized, that most the time was spent
on IPC. Increasing the calc_2d_hist size solved this problem.